Added Rectangle object logic and rendering

+++ app/src/main/java/com/cuseto/pong/model/Circle.java
@@ -4,9 +4,9 @@ package com.cuseto.pong.model;
  * Encapsulates the ball's position and radius
  */
 public record Circle (
-    int startPosX,
-    int startPosY,
-    int radius
+    double startPosX,
+    double startPosY,
+    double radius
 ) {
     public Circle {
         if (radius <= 0) {
